Remove map component and fix styling #32
