+++ screens/map-screen/map-screen.component.jsx
@@ -3,11 +3,24 @@ import { StyleSheet, View } from "react-native";
 
 import { RegionProvider } from "../../context/region.context";
 
-import Map from "../../components/map/map.component";
 import CampusToggle from "../../components/campus-toggle/campus-toggle.component";
+import MapView, { PROVIDER_GOOGLE } from "react-native-maps";
+import BuildingHighlights from "../../components/building-highlights/building-highlights.component";
+import SlidingPanel from "../../components/sliding-panel/sliding-panel.component";
 
 const MapScreen = () => {
   const [region, setRegion] = useState(null);
+  const [showAdditionalInfo, setShowAdditionalInfo] = useState(false);
+  const [tappedBuilding, setTappedBuilding] = useState("");
+
+  const onDisplayBuilding = (showAdditionalInfo, tappedBuilding) => {
+    setShowAdditionalInfo(showAdditionalInfo);
+    setTappedBuilding(tappedBuilding);
+  };
+
+  const onClosePanel = showAdditionalInfo => {
+    setShowAdditionalInfo(showAdditionalInfo);
+  };
 
   useEffect(() => {
     setRegion({
@@ -21,8 +34,30 @@ const MapScreen = () => {
   return (
     <RegionProvider value={{ region, setRegion }}>
       <View style={styles.container}>
-        <Map />
-        <CampusToggle />
+        <MapView
+          style={styles.map}
+          provider={PROVIDER_GOOGLE}
+          showsCompass={true}
+          showsBuildings={true}
+          showsUserLocation={true}
+          region={region}
+          // onRegionChangeComplete={region => setRegion(region)}
+        >
+          <BuildingHighlights
+            tappedBuilding={tappedBuilding}
+            showAdditionalInfo={showAdditionalInfo}
+            displayBuilding={onDisplayBuilding}
+          />
+        </MapView>
+
+        <SlidingPanel
+          tappedBuilding={tappedBuilding}
+          showAdditionalInfo={showAdditionalInfo}
+          closePanel={onClosePanel}
+        />
+        <View style={styles.campusToggle}>
+          <CampusToggle />
+        </View>
       </View>
     </RegionProvider>
   );
@@ -30,10 +65,16 @@ const MapScreen = () => {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center"
+    ...StyleSheet.absoluteFillObject,
+    justifyContent: "flex-end",
+    alignItems: "center"
+  },
+  map: {
+    ...StyleSheet.absoluteFillObject
+  },
+  campusToggle: {
+    position: "absolute",
+    bottom: 0
   }
 });
